Moved image related operations to separate Namespace
