+++ app/__init__.py
@@ -7,6 +7,7 @@ from app.main.controller.auth_controller import api as auth_ns
 from app.main.controller.issue_controller import api as issue_ns
 from app.main.controller.tag_controller import api as tag_ns
 from app.main.controller.user_controller import api as user_ns
+from app.main.controller.image_controller import api as image_ns
 
 blueprint = Blueprint('api', __name__)
 
@@ -21,3 +22,4 @@ api.add_namespace(auth_ns, path='/auth')
 api.add_namespace(post_ns, path='/post')
 api.add_namespace(tag_ns, path='/tag')
 api.add_namespace(issue_ns, path='/issues')
+api.add_namespace(image_ns, path='/image')
